feat: ajout de la vérification que la période n’est pas encore clôturée avant l’ajout ou la suppression d’un évaluateur dans EvaluatorController.php

+++ app/Http/Controllers/EvaluatorController.php
@@ -74,6 +74,12 @@ class EvaluatorController extends Controller
     {
         try {
             $type = $request->type;
+            $period = Period::query()->findOrFail($request->periodId);
+
+            if ($period->status == PeriodStatusEnum::STATUS_CLOSE->value) {
+                throw new \Exception("Vous n'avez plus le droit d'ajouter un evaluateur pour cette periode.");
+            }
+
             $ldapUser = $this->userLdapService->findUserByCuid($request->cuid);
 
             $exists = User::where('email', $ldapUser->email)->exists();
@@ -101,7 +107,6 @@ class EvaluatorController extends Controller
                 throw new \Exception("L'utilisateur est déjà enregistré en tant qu'évaluateur pour l'épreuve de {$type}.");
             }
 
-            $period = Period::query()->findOrFail($request->periodId);
 
             if ($period->status != PeriodStatusEnum::STATUS_DISPATCH->value && $type == EvaluatorTypeEnum::EVALUATOR_PRESELECTION->value) {
                 throw new \Exception("Vous n'avez plus le droit d'ajouter un evaluateur de PRESELECTION pour cette periode.");
@@ -143,9 +148,22 @@ class EvaluatorController extends Controller
      */
     public function destroy(string $id)
     {
-        $evaluator = Evaluator::query()
-            ->findOrFail($id);
-        $evaluator->delete();
+        try {
+            $evaluator = Evaluator::query()
+                ->findOrFail($id);
+
+            $period = Period::query()->findOrFail($evaluator->period_id);
+
+            if ($period->status != PeriodStatusEnum::STATUS_DISPATCH->value) {
+                throw new \Exception("Vous  n'avez pas le droit d'effacer cet evaluateur car le status de la periode est PRESELECTION.");
+            }
+
+            $evaluator->delete();
+        } catch (\Exception $e) {
+            throw  new HttpResponseException(
+                response: response()->json(['errors' => $e->getMessage()], 400)
+            );
+        }
     }
 
     public function evaluatorCandidacy(int $id): AnonymousResourceCollection
